requirements.txt
Django==1.4.3
Pillow==1.7.7
South==0.7.6
Werkzeug==0.8.3
django-classy-tags==0.3.4.1
django-debug-toolbar==0.9.4
django-extensions==0.9
django-filebrowser==3.5.1
django-grappelli==2.4.3
django-mptt==0.5.4
django-polymorphic==0.2
django-shop==0.1.1
django-shop-categories==0.8.beta.2
django-sitetree==0.9.1
django-tinymce==1.5.1b4
django-treeadmin==0.1
ipython==0.11
sorl-thumbnail==11.12
wsgiref==0.1.2
